+++ src/Symbolize/SymbolTable.hs
@@ -82,8 +82,8 @@ insertGlobal ba# = do
   -- Optimization: Look up without doing a CAS first
   -- If we can find anything, great!
   readTable <- IORef.readIORef gsymtab
-  let !key = calculateHash sipkey ba#
-  case lookup ba# sipkey readTable of
+  let !hash = calculateHash sipkey ba#
+  case lookup ba# hash readTable of
     Just ba -> pure ba
     Nothing -> do
       -- SAFETY: If the table IORef contested, 
@@ -92,19 +92,20 @@ insertGlobal ba# = do
       -- But finalization is idempotent, and only when a thread finally wins the Compare-and-Swap
       -- will its `weak` pointer be inserted (or alternatively another previously-inserted `ba` returned).
       -- So once this function returns, we can be sure we've returned a deduplicated ByteArray
-      !weak <- mkWeakSymbol ba# (removeGlobal key)
+      !weak <- mkWeakSymbol ba# (removeGlobal hash)
       IORef.atomicModifyIORef' gsymtab $ \table ->
-        case lookup ba# sipkey table of
+        case lookup ba# hash table of
           Just ba -> (table, ba)
           Nothing ->
-            (insert key weak table, ByteArray ba#)
+            (insert hash weak table, ByteArray ba#)
 
 lookupGlobal :: ByteArray# -> IO (Maybe ByteArray)
 {-# INLINE lookupGlobal #-}
 lookupGlobal ba# = do
   GlobalSymbolTable gsymtab sipkey <- globalSymbolTable
   table <- IORef.readIORef gsymtab
-  pure (lookup ba# sipkey table)
+  let hash = calculateHash sipkey ba#
+  pure (lookup ba# hash table)
 
 removeGlobal :: Hash -> IO ()
 {-# INLINE removeGlobal #-}
@@ -119,11 +120,10 @@ insert key weak (SymbolTable table) =
   let table' = IntMap.alter (Just . maybe (pure weak) (NonEmpty.cons weak)) (hashToInt key) table
    in SymbolTable table'
 
-lookup :: ByteArray# -> SipHash.SipKey -> SymbolTable -> Maybe ByteArray
+lookup :: ByteArray# -> Hash -> SymbolTable -> Maybe ByteArray
 {-# INLINE lookup #-}
-lookup ba# sipkey (SymbolTable table) = do
-  let !key = calculateHash sipkey ba#
-  weaks <- IntMap.lookup (hashToInt key) table
+lookup ba# hash (SymbolTable table) = do
+  weaks <- IntMap.lookup (hashToInt hash) table
   Foldable.find (\other -> other == ByteArray ba#) (aliveWeaks weaks)
 
 remove :: Hash -> SymbolTable -> SymbolTable